♻️ refactor: CommentSection props 정리 및 댓글 에러 처리 개선
