+++ src/app/(routers)/(board)/community/[id]/_components/CommentSection.tsx
@@ -2,38 +2,47 @@
 
 import type { Comment } from '../../types';
 
-import { useEffect, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useInfiniteScroll } from '@/hooks/useInfiniteScroll';
 
-import { useCreateComment, useDeleteComment } from '../../_api/communityQueries';
+import { useCreateComment, useDeleteComment, useGetComments } from '../../_api/communityQueries';
 import { CommentItem } from './CommentItem';
 
 interface CommentSectionProps {
   postId: number;
-  comments: Comment[];
   totalCount: number;
   userId: number | undefined;
-  isBusy?: boolean;
+  isPostDeleting?: boolean;
   onPendingChange?: (isPending: boolean) => void;
-  hasNextPage: boolean;
-  isFetchingNextPage: boolean;
-  fetchNextPage: () => void;
 }
 
 export function CommentSection({
   postId,
-  comments,
   totalCount,
   userId,
-  isBusy = false,
+  isPostDeleting = false,
   onPendingChange,
-  hasNextPage,
-  isFetchingNextPage,
-  fetchNextPage,
 }: CommentSectionProps) {
+  const {
+    data: commentsData,
+    isError,
+    isFetching,
+    refetch,
+    fetchNextPage,
+    hasNextPage,
+    isFetchingNextPage,
+  } = useGetComments(postId);
+
+  const comments: Comment[] = useMemo(() => {
+    const all = (commentsData?.pages ?? []).flatMap((page) => page.comments);
+    return Array.from(new Map(all.map((c) => [c.id, c])).values());
+  }, [commentsData]);
+
   const [inputValue, setInputValue] = useState('');
   const { mutate: createComment, isPending: isCreating } = useCreateComment(postId);
   const { mutate: deleteComment, isPending: isDeleting } = useDeleteComment(postId);
+
+  const isBusy = isPostDeleting || isFetching;
 
   useEffect(() => {
     onPendingChange?.(isCreating || isDeleting);
@@ -80,25 +89,40 @@ export function CommentSection({
         <button
           type="button"
           onClick={handleSubmit}
-          disabled={!inputValue.trim() || isBusy}
+          disabled={!inputValue.trim() || isBusy || isCreating}
           className="font-sm-semibold md:font-base-semibold w-16 shrink-0 rounded-full bg-orange-500 py-2.5 text-white disabled:cursor-not-allowed disabled:bg-gray-300 md:w-20 md:py-3"
         >
           등록
         </button>
       </div>
 
-      <ul className="flex flex-col gap-8 md:gap-10">
-        {comments.map((comment) => (
-          <CommentItem
-            key={comment.id}
-            comment={comment}
-            isMyComment={comment.userId === userId}
-            onDelete={deleteComment}
-            isDeleting={isBusy}
-          />
-        ))}
-      </ul>
-      <div ref={observerRef} className="h-4" />
+      {isError ? (
+        <div className="flex flex-col items-center gap-2 py-6 text-center">
+          <p className="font-sm-regular text-gray-500">댓글을 불러오지 못했습니다.</p>
+          <button
+            type="button"
+            onClick={() => void refetch()}
+            className="font-sm-medium text-orange-500 underline"
+          >
+            다시 시도
+          </button>
+        </div>
+      ) : (
+        <>
+          <ul className="flex flex-col gap-8 md:gap-10">
+            {comments.map((comment) => (
+              <CommentItem
+                key={comment.id}
+                comment={comment}
+                isMyComment={comment.userId === userId}
+                onDelete={deleteComment}
+                isDeleting={isBusy}
+              />
+            ))}
+          </ul>
+          <div ref={observerRef} className="h-4" />
+        </>
+      )}
     </div>
   );
 }
